requirements.txt
Flask
Flask-WTF
gunicorn
supabase-py
python-dotenv